Admin button & refactoring React components
