+++ frontend/src/Components/LoginButton.jsx
@@ -1,0 +1,9 @@
+function LoginButton(){
+    return(
+        <>
+            <button className="btn btn-outline-dark ms-auto">Admin Login</button>
+        </>
+    )
+}
+
+export default LoginButton;
